Revert "Allow instance-id to auto-generate typings, separate internal vs external APIs (#965)"
This reverts commit 289561075f624e408a567f769eed6bff1f8b4f3d.
