+++ gulpfile.js
@@ -52,7 +52,7 @@ var paths = {
 
   build: 'lib/',
 
-  curatedTypings: ['src/*.d.ts', '!src/instance-id.d.ts'],
+  curatedTypings: ['src/*.d.ts'],
 };
 
 const TEMPORARY_TYPING_EXCLUDES = [
@@ -63,6 +63,7 @@ const TEMPORARY_TYPING_EXCLUDES = [
   '!lib/auth/*.d.ts',
   '!lib/database/*.d.ts',
   '!lib/firestore/*.d.ts',
+  '!lib/instance-id/*.d.ts',
   '!lib/machine-learning/*.d.ts',
   '!lib/messaging/*.d.ts',
   '!lib/project-management/*.d.ts',
