+++ src/firebase-app.ts
@@ -30,7 +30,6 @@ import { DatabaseService } from './database/database';
 import { Firestore } from '@google-cloud/firestore';
 import { FirestoreService } from './firestore/firestore';
 import { InstanceId } from './instance-id/instance-id';
-import { InstanceIdImpl } from './instance-id/instance-id-internal';
 
 import { ProjectManagement } from './project-management/project-management';
 import { SecurityRules } from './security-rules/security-rules';
@@ -353,7 +352,7 @@ export class FirebaseApp {
    */
   public instanceId(): InstanceId {
     return this.ensureService_('iid', () => {
-      const iidService: typeof InstanceIdImpl = require('./instance-id/instance-id-internal').InstanceIdImpl;
+      const iidService: typeof InstanceId = require('./instance-id/instance-id').InstanceId;
       return new iidService(this);
     });
   }
